requirements.txt
django==1.10.2
django-crispy-forms==1.6.0
django-picklefield==0.3.2